created axiosWithAuth. created utils folder and moved authHelper
