+++ buddy-app/components/AuthStack.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import Dashboard from "./Dashboard";
 import SignIn from "./SignIn";
-import { isSignedIn } from "../authHelper";
+import { isSignedIn } from "../utils/authHelper";
 import { connect } from "react-redux";
 import { isLoadingPage } from "../actions/buddyActions";
 const AuthStack = props => {
