+++ buddy-app/utils/axiosWithAuth.js
@@ -1,0 +1,11 @@
+import axios from "axios";
+import { getToken } from "./authHelper";
+
+export const axiosWithAuth = () => {
+  let token = getToken();
+  return axios.create({
+    headers: {
+      Authorization: token
+    }
+  });
+};
